refactor(server): extract Hono-coupled utilities to backend-neutral modules
Move TUI control queue, fence sync helpers, workspace routing helpers,
and the UI-asset Effect into packages/opencode/src/server/shared so the
HttpApi backend no longer reaches into Hono route files for them. The
Hono adapters (FenceMiddleware, WorkspaceRouterMiddleware, UIRoutes,
TuiRoutes) stay where they are and re-import from the neutral modules.

Pure relocation; no behavior change.
